[FAB-8075] Peer Deliver client doesn't failover to OSN
If the peer is connected to a specific ordering node, and no
blocks are received within a timely manner, when that ordering
node loses connection - the peer gives up connecting to a different
ordering node.

This is because the total elapsed time includes the time waiting
on the Recv() and it instead should only include the time
that is spent while sleeping because of the exponential backoff.

Change-Id: I4a75dad1f1fa2facf70498a3d9692598b89e8a44
Signed-off-by: yacovm <yacovm@il.ibm.com>