[FAB-7886] Exclude non .json ext in deploy metadata
When using Golang CC, the presence of a *.txt file (containing multiple
json indexes) does not cause any issues in creating indexes from valid
index definition files, but while using Node CC, the peer crashes with
an error.

Add a check in the ValidateMetadataFile to verify the file extension
is ".json".  Add a return value to indicate if the file should be included
since returning an error would cause the peer to crash.

Change-Id: Iceeca2d6683490eda9330becbe95a8a4ecedb5c0
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>