[FAB-8336] Add create or update msg for CouchDB index
Problem Statement:

If the provided index definitions have more than one (lets say N)
index with same DesignDoc and indexName but with potentially different indexes,
the log will contain N messages like `created CouchDB index in the channel...`
but in reality only 1 index will be created in CouchDB. Fabric goes through
each file in alphabetic order, creating and (quietly) replacing the index
created in last round with the same DDoc and indexName, so only the last index
remains in CouchDB.

Solution:

Part of the confusion can be prevented by updating the message when an index
is created or updated in CouchDB.  The create index command responds with
metadata that can be captured and provide a more detailed message and can
determine if the index has been created or updated.  This can be accomplished
with the existing http request.

Change-Id: I84674ceed54dadba9470d768a028a6acb3b237f6
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>