[FAB-8183] Remove old configtxlator doc
The new command doc layout makes the old configtxlator document
superfluous.  However, there was still some content in the old doc that
was useful to preserve, and has not been moved to the command doc.

The old configtxlator doc also contained a number of examples, which are
being obsoleted by the newer reconfiguring tutorials.  Because
reconfiguring your network is a different task from simply using
configtxgen, this content would not be appropriate to port.  Therefore
the bulk of it is simply deleted.

Change-Id: I1a2bae5f4f89ee647ad7b00082c638f4958f1e44
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>
Signed-off-by: David Enyeart <enyeart@us.ibm.com>