[FAB-8590] Fix upgrade doc to reflect sample
Per https://gerrit.hyperledger.org/r/c/18535/
and https://gerrit.hyperledger.org/r/c/18539/

The first-network sample is going to use volumes for the ledger storage
by default.  This makes the need of the second compose file and the '-p'
option no longer necessary.  This CR updates the doc to reflect this.

Change-Id: I495e5ffaaf82e4a0f72d21b24328ad4429258137
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>