[FAB-8688] Gossip onConnect() references wrong field
In gossip, when a peer p contacts peer q and tries to connect
to it, while it's already connected to it - q closes the current
connection with p and replaces it with q.

However, the code referenced the identity field instead of the
pki-id field and thus it always replaced the connection,
without closing the previous connection.

Change-Id: I021ce246ce3735aa9e3451f20590fd119737a310
Signed-off-by: yacovm <yacovm@il.ibm.com>