[FAB-8844] correct handling of launch errors
An error during launch is incorrectly handled by partial deregistration.
This error maybe benign and ignored by system, thus causing the
deregistration to leave it in an inconsitent state where it believes
the chaincode is not registered (and running) when it actually is.
This causes subsequent invoke to attempt to restart.

The fix is to have errors handled at the right decision point with existing
error mechanism naturally handling them..

Change-Id: I9c655a5355991b2426b88986aa7da1f7145ba15e
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>