[FAB-8870] Return concrete types from constructors
The contructors for the gRPC client and server were
actually returning interfaces rather than
concrete types.  As there was really no use for
the interfaces in the package, the interfaces
were removed and concrete types are now public and
returned from the constructors.

Additionally, some cleanup on the types as well
to remove redundant fields.

Change-Id: I1c82f77aab3419fcde4f04e6a083428b85086bae
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>