[FAB-8976] Move fixtures closer to tests
Test fixtures should typically be placed in `testdata` directories to
prevent the go tooling from operating on them by default. There are a
number of these fixtures that live in the incorrectly named top level
`test` directory.

Instead of keeping fixtures in one location, this commit moves the
assets closer to the tests that use them and makes the necessary path
modifications to reference them.

Tests that were relying on fixtures to live on the go path, were
modified to explicitly set the gopath to the testdata directory.

Change-Id: I5dbc0cb5af41c63289aea925b08fea1584ca598f
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>