[FAB-9256] Change DEFAULT MSPID to SampleOrg
Our users are frequently confused because the sample configurations
set to the sample organization's MSP ID of 'DEFAULT'.  This leads to
very misleading error messages, and it is not at all clear that
'DEFAULT' should be changed to match their organization.

Unfortunately, many of our 'unit tests' are actually integration tests
which hardcode the string 'DEFAULT', so this CR (which should only be a
few lines long) actually cascades through much of the code.  These tests
should be updated to remove their external dependency on the sample
configuration, but, that is outside the scope of this CR.

Change-Id: Icf56924629b69400e0d70991a6ecfb3e831b0641
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>