[FAB-9104] Improve core/comm tests
There were a number of flaky tests in
the core/comm package after upgrading the
gRPC package.  Many seemed to be due to
running tests in parallel while others
were just due to changes in library
behavior.

Fixes flaky TestNewConnection test as well as
eliminating race conditions for the tests

Change-Id: If2465ab3332408d7c7d7b87abe14ec4ac4393bc8
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>