[FAB-9389] Fix deliver client crash at abrupt restart
When the deliver client is started and then immediately stopped
for a certain channel, the goroutine that runs the block provider
accesses a nil pointer because the stop operation deletes an entry
from the map.

Moreover, it accesses the map without a lock which is a problem
since entries can be removed from the map and added to the map.

Change-Id: Icec5b7e4398b044af01e0c7585ea8e48e7cb0abf
Signed-off-by: yacovm <yacovm@il.ibm.com>