[FAB-9174] Fix peer crash due to log contention
In prior releases, the version of go-logging
vendored by Fabric was directly modified to
add locking around getting and setting log
levels.  In the current src tree, the
unmodified version of go-logging is used.

The issue from FAB-9174 should not occur if
logging is actually initialized prior to
be used in other modules.

Change-Id: I0be40796b1b6f5e9ab54f8f50a1858724a1abe47
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>