[FAB-8551] Tool image binaries don't show version
Version is set via the use of build time flags
via the GO_LDFLAGS variable.  The issue here
was that the peer uses different flags than the
other binaries.  While it would have been possible
to set the flags appropriately, it would have added
a few more entries in the Makefile.

Instead, this change leverages the multi-stage
build feature of modern versions of Docker.  This
change is limited to the fabric-tools container,
but hopefully once this is merged we can explore
doing this for other images as well.

Change-Id: Ief4390283303b71514f39fd0671df2290bfd0191
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>