[FAB-9473] Always enforce validation on initial cc
Currently, if a tx does not modify any data, all validation of that tx
is skipped.  This is usually not the application writer's intent, and
instead, the initially called chaincode should always have its EP
enforced, regardless of whether writes are made.

This also fixes the issue of enforcing endorsments for chaincode events.
Currently, only the initially invoked chaincode may include events in
the proposal results.  Therefore any chaincode events will be verified
using this EP.

Change-Id: I3c98e54be2ba4a53d26c8c13c0e243c6871aad34
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>