[FAB-9478] Principal set filtering based on role
This change set implements filtering of principal sets based on
their role and MSP ID.

This is needed for collection support for service discovery:
When a chaincode has several possible endorsement combinations
that each combination is a principal set, but the user wants
to perform a collection read/write - then the response should
contain only principal sets which are "authorized" by the
collection config.

For this - we need a way to filter out principal sets that
there exists a principal among them which isn't "authorized"
by the collection config.

This change set implements this filtering while taking into account
the different MSP roles, for instance- if the endorsement policy
has Peer roles in it, or OU roles, but the collection config has members,
then it would still work because a peer is also a member (but not vice versa)

Change-Id: I5b4c1a031b81e78ecdf535ae363a4c45a8a44967
Signed-off-by: yacovm <yacovm@il.ibm.com>