[FAB-9610] Unsupported fields appear supported
There are two v0.5/v0.6 fields in the current protos which have exactly
zero affect on the operation of Fabric v1.0+.  This is confusing enough
for our users, but we additionally provide a mechanism to set one of the
fields through the CLI (despite its meaninglessness).  This CR remedies
this situation by simply removing the proto definitions and flag.

Change-Id: Ic2cecb63c3d1ae9b3125ac5dbb26ffd02f89c7de
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>