[FAB-9745] close txsim ASAP once the tx is simu.
This CR closes txsim once the tx is simulated and the rwset is
collected.

txsim object acquires a shared lock on the stateDB which restrict block
commits, i.e., commiting valid write-sets to the stateDB. Hence, we should
close the txsim as soon as the tx is simulated and the rwset is collected.
Definitely before gossip dissemination, we must release the shared lock.

Change-Id: Ic1992b62596f131e666b447c32c5a34dae3f71ec
Signed-off-by: senthil <cendhu@gmail.com>