FAB-9751 Remove flaky logging concurrency test
The logging infrastructure is set to be replaced soon.  There is a test
which attempted to verify that the concurrency hacks were left in place
in the vendored package.  However, this vendored package has been
updated and the hacks removed, so this test is actually a false-negative
when it passes.  Removing it.

Change-Id: I6f96574ca03f0938a15e1a4a3adeca9910c42a9b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>