FAB-9295 Enforce event chaincode_id matches action
There was a change to enforce that the the initially invoked chaincode
is always validated.  This is to ensure that the chaincode event is
always validated.

If generated by a peer, the chaincode event's chaincode_id field will
always match the chaincode_id field of the ChaincodeAction.  However, a
custom crafted transaction could create a mismatch here, on a chaincode
with a weak validation policy.  So, we should enforce that the ids of
the event and the action match.  This CR does just that.

Change-Id: Ic4a55f075c175983d221e67df009a1a28f046266
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>