[FAB-9982] Service Discovery collection support
This change set:

Implements filtering for collections, for a single chaincode call.
cc2cc calls will be implemented in subsequent change sets.
If the given collections in the query don't exist, an error is returned.
If all collections exist, then combinations of principals that are returned
by service discovery are filtered such that all principals are found
in the collection config.

Change-Id: I65bb159ff8c0a4d5474fcff8b7b5e5ecc16035b5
Signed-off-by: yacovm <yacovm@il.ibm.com>