[FAB-10023] Principal set merging
This change set implements merging of principal sets sets.
More specifically, if 2 or more sets of principal sets are given,
such that each principal set set - satisfies an endorsement policy,
we are able to compute sets of principal sets such that each principal set
satisfies all of the endorsement policies.

This is to be used for cc2cc queries for service discovery.

Change-Id: I4b6e63f07b1b5832667475b7876df754007d9c2f
Signed-off-by: yacovm <yacovm@il.ibm.com>