[FAB-10036] Ledger: Metadata - Commit-path enhancements
This CR enhances the ledger validation and commit code for
processing metadata writes in the read-write set.
For each key that is present in the writes in the read-write set,
the final value and metadata that should be committed to the
state-db is derived as follows
 - Take the value from the highest valid transaction (Tv)
   that updates the value
 - Take the metadata from the highest valid transaction (Tm)
   that updates the metadata
 - Construct the the final tuple <value, metadata, version> for the key
   as <value in Tv, metadata in Tm, max(version of Tv, version of Tm)>
 - Finally, if Tv indicates delete of the key,
   the  final tuple also indicates the delete of the key

For more details on the semantics, see the associated Jira.

Change-Id: Icae615b0972fe985f8e86147cbcfb0bf0a06266c
Signed-off-by: manish <manish.sethi@gmail.com>
Signed-off-by: David Enyeart <enyeart@us.ibm.com>