[FAB-10076] Peer runner fails to list instantiated cc
The peer runner integration test currently fails to get the
instantiated chaincodes. This is due to the new pluggable
validation and the "GetInstantiatedChaincodes" policy not
existing at the correct scope. This CR adds the default VSCC
plugin information to core.yaml for the peer integration runner
and properly sets the default policy for "GetInstantiatedChaincodes"
to be channel-scoped instead of peer-scoped.

Change-Id: I763b3d84ebefc3af6a3c0dc5e841ae9999312969
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>