FAB-10121 Remove org scoped profile from struct
The notion of tying an organization to a specific profile does not make
sense.  An organization may be part of many profiles.  In fact, a
profile is principally defined by which organizations are a part of it,
not the other way around.

There is already a 'world' scoped profile for orderer and for peer
channel.  This CR simply removes the organization scoped one which did
not appear to be doing anything.

Change-Id: Idc5ed84eacd1773550bdc534e7497f5d7f2d9797
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>