[FAB-10051] Pluggable endorsement and validation e2e
This change set adds integration tests for pluggable endorsement
and validation.

This is actually the e2e but with a slight twist:
The endorsement and validation plugins that are used are actually
not the builtin, compiled ones - but .so plugins that are compiled
at the start of the test.

To ensure that the plugins are run and not the builtin version,
the plugin implementation writes to the file system a file
and the test checks that the file exists after the peers joined the channel.

Change-Id: I69027566f78560ffe9afc03aa9f3f1ac7bdbc6e5
Signed-off-by: yacovm <yacovm@il.ibm.com>