[FAB-10153] Continue instead of erroring- endorserState
The endorser's private data support assumes there is simulation
results if the iterator claims the iteration hasn't ended and
if there aren't any - it returns an error.

I think it's better be safe than sorry and just skip this iteration
and continue to the next one, and in the worst case - the user
anyway needs to inspect the results from this method.

Change-Id: I765057f6f4d231eceef55232b3831f5925a17dfa
Signed-off-by: yacovm <yacovm@il.ibm.com>