[FAB-9779] Cleanup CLI invoke error processing
This CR cleans up and updates the peer CLI's invoke error processing
logic. Given the new ability to define an endorser plugin that may
endorse a chaincode response with any status code that it desires,
the CLI now logs success/returns failure based on whether the proposal
contains an endorsement and not whether the status is greater than the
shim.ERRORTHRESHOLD. This CR also removes duplicated code and updates
errors to use the vendored errors package.

Change-Id: I6afc12e0204ea1421ce6c3eb8480070cf4226745
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>