FAB-10825 Make platform latent dependency explicit
Many of the assorted 'utils' functions actually reach into the
core/chaincode/platforms package to determine whether packages are
valid.  This historically has all been done in a source-coupled way,
pulling peer details into places it does not belong.  Although these
'util' functions should be refactored and removed, for the time being,
we can at least make the dependency on the platforms package explicit,
but requiring that an instance be passed into these util functions.

This has quite the ripple throughout the code, but the CR itself should
be trivial to review.

Change-Id: I0cc36e2f307474ddba7f6d20028a46f3aa94faf5
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>