[FAB-10222] Honor maximum peers to send if 0
When the collection config specifies max peers, it might be 0.

Currently, gossip assumes that if the max peers is 0 then it was
omitted, and just sets the max peers to all the known peers.

Since collection membership can change in v1.2, we should be
conservative and enforce 0 max peers as 0.

Change-Id: Idef84b6d5d88a719c737aef05b65a724a3a38468
Signed-off-by: yacovm <yacovm@il.ibm.com>