[FAB-10238] Filter collections on peers, not principals
The collection support in the discovery service supports collections by
filtering out principal sets that aren't authorized by the principals
derived from the policy in the collection config.

This is problematic in cases when the collection config is defined
with principals which have no obvious IsA relation with the
endorsement policy:
Consider a case where the endorsement policy is defined with principals
of Org1.peer or and org2.peer, and the collection is
defined with OUs of Org.

Any principal based policy can't accept nor reject such a scenario,
since a peer role is based on an OU that is hidden from the principal
itself.

This change set moves the filtering to the peer identities instead of
filtering the principal sets of the combinations of the endorsement
policy.

It also raises the code coverage from 97% to 99%

Change-Id: I6d2f9f1e735f667fcffe99c3d3b6bfe6a6ed8e98
Signed-off-by: yacovm <yacovm@il.ibm.com>