FAB-10975 Split Execute into Execute/ExecuteInit
Presently, the Execute function can either take a
'ChaincodeInvocationSpec' or a 'ChaincodeDeploymentSpec', and the path
switches in a few different places depending on which it is.  This makes
the code complicated and difficult to understand, and in the interest of
removing the chaincode deployment spec entirely, this code path should
be isolated.

Change-Id: I1f54684dba5877b5c629a59d344803f6423cc3da
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>