[FAB-10403] Fix in txmgr related to  purgemgr
This CR
- Moves one function call outside the `defer`statement
This call is on function that clears off the used purged entries
from purge schedule and should not get called in the case of an
error while applying updates to the statedb

- Makes the validation logic to wait for a background job
`PrepareForExpiringKeys` to finish. This background job is
launched after the commit of a block in an attempt to pre-fetch
and prepare the data needed for purging the eligible
keys along with the next block commit.
This wait notion is already in place as part of
'DeleteExpiredAndUpdateBookkeeping' function internally.
However, this needs to be externalized because both validation logic
and 'PrepareForExpiringKeys' use the version cache. Else, validation
will reload the cache and slowdown the `PrepareForExpiringKeys` (in
case, this was not completed by the time next block arrives).
Extending the version cache logic would be ideal so that the caller
is direcly given off the cache to work on independently but that
may require some refactoring and can be taken up post 1.2. Will
open a separate Jira for this.

Change-Id: I09a66fefce7984372e85b29f99a032caff523cf3
Signed-off-by: manish <manish.sethi@gmail.com>