FAB-10979 PackageProvider provides code package
The PackageProvider presently returns a chaincode deployment spec.
However, this is far more information than the chaincode runtime
launcher actually needs, as it already knows the chaincode version, type
etc.  So, this CR adds a small method to the cc fs provider to return
only the package bytes and not the rest of the cds and removes the
associated complexity from the chaincode runtime launcher and tests.

Change-Id: Id5ed31ed1fa5030bbb530c55e7c65f6f9a8eccc2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>