FAB-10521 Block-cutter should refetch config
The block-cutter is currently instantiated with a reference to the
orderer channel config.  Once upon a time, each config element was its
own mutable state machine, so this worked fine for v1.0.x.  However, in
v1.1, the individual mutable configs were transformed to be a single
immutable config bundle.  When the config changes, an entirely new
immutable bundle is created and atomically swapped out.

Because of this, the block-cutter only retains the reference to the
initial immutable blob.  It needs to be fixed to pull a fresh reference
of the config through the atomic dereference at every block.  This CR
does just that.

Change-Id: I6be1b65cb8d8cc86d5e47da906a127da6651354e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>