FAB-10997 Add ChaincodeDefinition function to LSCC
Currently, the endorser and chaincode package get chaincode definitions
by invoking the LSCC.  This creates a needless round of marshaling and
unmarshaling and other chaincode overhead, plus makes the function
signatures large and hard to read.  This CR adds a new
ChaincodeDefinition method to LSCC which simply accepts a txsim and a
chaincode name.  With these two things, the necessary reads are attached
to the transaction while bypassing the rest of the chaincode overhead.

Change-Id: I1fcf2c9a71f4713bfd391138fc0b1458e84fa32f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>