[FAB-10574] txmgr: rel lock only if it was acquir.
In the transaction manager, we need to release the write lock
only if the lock was acquired. When an error occurs before
acquiring a lock and if we try to release the unacquired lock
in the defer statement, we would end up printing the wrong
panic message (i.e., fatal error: sync: unlock of unlocked
mutex) which would be less helpful during troubleshooting.

Change-Id: I7b69f6a229a6edce374a960d1467a09057bfe25e
Signed-off-by: senthil <cendhu@gmail.com>