FAB-10577 User supplied code-package has priority
There is still cruft from v0.5/v0.6 in the chaincode package.  If a user
supplies a code-package with an instantiate (deploy or upgrade), this
could cause the chaincode package to erroneously ignore the package on
the local filesystem.

This CR addresses the problem by retrieving the CDS from the filesystem
first, nils out the code-package, then copies the user-init-args into
it.  The correct fix is to remove the notion of CDS from the chaincode
package, but, that is a much longer road which should be done in the
future.

Change-Id: Icd418f8974c1afc07730b727775f6783763feb9d
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>