[FAB-10576]: Check if puller got valid pvt rwset
While puller retrieves missing pvt data, whenever it's from ledger or
from transient store there is might be case it has failed due to some
error. In that case we need to add proper log message to the body of the
puller createResponse function and skip to the next digest. This commit
add such check for rwSets is not nil to prevent peer from crashing with
panic.

Change-Id: Ie6bb6e85520eac6992872da862d24c30dc2ffe34
Signed-off-by: Artem Barger <bartem@il.ibm.com>