[FAB-8557] Fix overwriting txid in block index
This CR fixes the current behavior of block store index.
In the current implemetation, the indexes that maintain
different pieces of information by txid are overwritten
by a duplicate txid (if any). This CR reverses this behavior
and keeps the first appearance of txid in the index. The future
duplicate txids will not overwrite it.

Though, either of these behaviors (keeping the first tx or
keeping the last tx) are not fully justified, primarily because
the problem itself is a paradox - in the sense that the ids
(that are supposed to be unique by definition) are duplicated.
However, the justification of moving from the current behavior
to the proposed behavior is that, its easy for someone to replay
the transaction or simply use an existing txid in a bogus transaction
just to exploit the current behavior of overwriting the block storage
index and preventing the legitimate user to query about the status
of their transactions.

Change-Id: I3b81ae61c756ef78253b58a94644778716fb0e16
Signed-off-by: manish <manish.sethi@gmail.com>