FAB-11037 Remove InvocationSpec from chaincode pkg
This CR finally completes removing the legacy data "ChaincodeSpec"
structures from the non-shim related pieces of the chaincode package.
To maintain compatibility with the existing chaincode images, the
ChaincodeSpec still exists in the handler path, and to maintain
compatibility with the legacy lifecycle, a single deployment spec
acceing execute method is available.  Both of these should be removed in
time.

Change-Id: I5659695a55e0947a1ccec23721192a4c88177b38
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>