FAB-9641 add CommitSHA metadata
Added metadata.CommitSHA
Made metadata.Version a const that would be
modified with each MMP release and allow for
binaries to be built that could be released directly
following successful pipeline according to the
stable image strategy.

This also has the side-effect of addressing the issue
raised in FAB-8559 that the fabric-ccenv image might not
be found in CI testing under certain circumstances.

rolling back last two commits... this one works.

Change-Id: Id2f4a172f181c395a2a1b8864c51d1431c3c3230
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>