[FAB-10617] Add writeset validation check during commit
Fabric version 1.1, missed one check in the couchdb key-values
during simulation. Couchdb does not allow an underscore as a
first character in the key or first-level fields name in the
JSON value. The implication of this is that during commit, when
the write-set is applied to the couchdb, the peer panic happens.

The problem is severe because this brings network at halt. This
is because, upon peer startup, peer again tries to process the
last block and hence gets into an endless cycle of panic and
restart.

This CR adds a check in the commit path for the validity of the
write-set and marks the transaction as invalid.

Change-Id: I36070bdba0a5012d6e99bc311e84be13863e7d8b
(cherry picked from commit 5d8a35eb1039c99dfed377cccb4f5e3e234a860e)
Signed-off-by: manish <manish.sethi@gmail.com>