[FAB-10559] fix iter in GetPrivateDataQueryResult
While creating an iterator for GetPrivateDataQueryResult
in the chaincode shim, we incorrectly assign value to
fields in the iterator object. For e.g., we assign
channelID to the txid field and txid to the channelID
field.

As a result, the peer is unable to get the txcontext while
processing the iter.next() and iter.close() call from
the shim.

This CR creates the correct iterator object and pass it
to the chaincode.

Change-Id: I360ad3e0768dd6f2bef8439d06d3967225f88a12
Signed-off-by: senthil <cendhu@gmail.com>