[FAB-10655]Problematic client conn stops eventhub evts
If a client connection hangs, the eventhub’s flow of events
completely stops until the problematic client connection decides
to return (likely after a grpc timeout). This CR adds a timeout
to the Send calls on the GRPC stream to ensure the eventhub is
not at the mercy of the client connection releasing the connection
before it can continue sending events. It also cleans up some code
and removes a large chunk of dead code, including the ability to
send chaincode events, which has not been used by Fabric at least
beginning with v1.0 if not earlier.

Change-Id: Idd86c1cec7d307becbd1a47a15c19f65eb0c64f1
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>