[FAB-10714] fix data race on grpc.EnableTracing
Setting the global variable inside of a connection factory does not
provide any guarantees that some other component in the process has not
already created and/or established gRPC sessions. This has resulted in
data races in the tests.

Moving the code to the peer initialization where it should be safer to
set.

Change-Id: Ifa6c87c88d1336ceea72755db85de19774ba8082
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>