FAB-10799 Cancel iterator with context
Presently, if in the middle of a deliver request, the server will block
indefinitely waiting for block creation or internal error before
returning.  This means that if the client requests blocks forever, then
hangs up, the client resources, including the ledger iterator and the
gRPC stack will not be released until a block is produced.

Change-Id: Ie8d9069ed4ad4dca436e388bebc23012b83eb7e9
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>