[FAB-10747] Fix endorsedAt block in transientstore
Transient store uses simulation block height to determine
purging schedule for orphaned (uncommitted) private data in transient
store.

There was a problem where endorser was not passing the simulation
block height to transient store persist function, causing 0 to be
used as the simulation height. The impact was that all private data
was getting purged from transient store at each purge cycle,
including private data for recent in-flight transactions.

This commit passes the simulation height to the transient
store, and removes SimulationBlkHt from TxSimulationResults, as
it was never set or used as initially intended.

Change-Id: I28523d503c1233f2c0221fb8efcc9890f89fae3a
Signed-off-by: David Enyeart <enyeart@us.ibm.com>