[FAB-10924] data race during event handler cleanup
When sending a message to a handler fails due to session timeout or some
other transient failure, the channel range loop in the start method will
cleanup the handler from the event processor.

When the event server stream associated with a handler encounters an
error, the server will cleanup the handler from the event processor.

When the two things happen concurrently, two go routines will
concurrently execute cleanup. This results in a concurrent map
modification on the handler's interestedEvents map.

In context, it doesn't appear that the map removal is needed as the
handler's interests are being purged. Removing this processing resolves
the race.

Change-Id: I2383482c25a08dfa3045d696b7c20551fd9bc8ba
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>