[FAB-10857] Extract discovery endorsement filtering
This change set extracts the logic that deals with discovery
endorsement peer filtering into a separate function, to be used
by the discovery service core for filtering peers according
to collections.

As a result, the code that pre-filters principal sets according
to MSP IDs is removed and the filtering is done implicitly
at the time of layout construction (if no peers with the chaincode
are installed for the given principal, the layout is filtered out in
the isLayoutSatisfied() function).
This removes the customly crafted MSPOfPrincipal method from existence.

Change-Id: I89a3b5142a6893d84a8faf0b65e2fed36829bb7d
Signed-off-by: yacovm <yacovm@il.ibm.com>