[FAB-10970] Make connection refusal more lenient
If the gossip communication layer is called to send a message to a peer
with a given PKI-ID, but after the handshake it discovers the remote peer
has a different PKI-ID than what was expected, it aborts the connection.

This is prolematic for cases where a peer has renewed its certificate, as
the PKI-ID which is a hash on the certificate, won't be the same - and
as a result, the reincarnated peer would be isolated.

This change set makes the connection be aborted only if the peer is
from a different organization.

Change-Id: I8e13dbce90a9df86eb40912f6e8105e8f19ef776
Signed-off-by: yacovm <yacovm@il.ibm.com>