[FAB-10302] Don't use protobuf as map keys
PvtDataDigest is a protobuf message that is used in the gossip private data
implementation.

The latest version of protobuf generates functions to protobuf structs,
which prevents using protobuf messages as map keys.

This change set moves the code to use an equivalent struct
that has the same fields but without functions.

Change-Id: Icdfdffb98625cb7ace37d6c0227906971e05fcc7
Signed-off-by: yacovm <yacovm@il.ibm.com>