[FAB-11414] Revert FAB-11377
This reverts commit 0e845ce1ce550f8fcaea12e13cb4a7cc736a5af2.

The -r / --no-run-if-empty flags are not supported by the BSD version of
xargs and causes the docker-clean targets to fail on MacOS.  Given a
great deal (if not a majority) of development is done on a Mac, this is
a problem.

The original change was intended to address a "failed" docker clean
target on Linux but it appears the target never failed as the recipe
command started with a hyphen (-).  When a target starts with a hyphen,
its exit status is ignored. I can see how the target would be noisy but
it shouldn't have failed.

Change-Id: I438e55cc761ff9ce8795d914f885a2e90165f78c
Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>